Move Mood to other class (interface class getting too big)
