+++ Mood_Busters/Mood_Busters.csproj
@@ -88,6 +88,7 @@
     <Compile Include="IErrorHandler.cs" />
     <Compile Include="IRecognitionApi.cs" />
     <Compile Include="Key.cs" />
+    <Compile Include="Mood.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StringConst.Designer.cs">
